Lint: ruff --fix + format across src, tests and scripts
Also adds .cursorrules with accumulated lessons and the task scratchpad.
No behavioural changes; full suite and smoke pipeline re-verified green.

+++ src/core/io_utils.py
@@ -179,7 +179,7 @@ def human_duration(seconds: float) -> str:
     seconds = max(0.0, float(seconds))
     if seconds < 60:
         return f"{seconds:.1f}s"
-    total = int(round(seconds))
+    total = round(seconds)
     days, rem = divmod(total, 86400)
     hours, rem = divmod(rem, 3600)
     minutes, secs = divmod(rem, 60)
